Moving Windows scheduler parallel to other custom schedulers
