+++ stlab/channel.hpp
@@ -17,6 +17,14 @@
 
 #include <stlab/future.hpp>
 
+#ifdef max
+#undef max
+#endif
+
+#ifdef min
+#undef min
+#endif
+
 /**************************************************************************************************/
 
 namespace stlab {
