+++ stlab/future.hpp
@@ -29,6 +29,8 @@
 #include <ppapi/cpp/module.h>
 #include <ppapi/cpp/core.h>
 #include <ppapi/cpp/completion_callback.h>
+#elif STLAB_TASK_SYSTEM == STLAB_TASK_SYSTEM_WINDOWS
+#include <Windows.h>
 #elif STLAB_TASK_SYSTEM == STLAB_TASK_SYSTEM_PORTABLE
 // REVISIT (sparent) : for testing only
 #if 0 && __APPLE__
@@ -1452,8 +1454,6 @@ using main_scheduler = default_scheduler;
     || (STLAB_TASK_SYSTEM == STLAB_TASK_SYSTEM_PNACL) \
     || (STLAB_TASK_SYSTEM == STLAB_TASK_SYSTEM_WINDOWS)
 
-// REVISIT (sparent): Since we are using the default threadpool, can the windows default scheduler
-// just be inlined here?
 
 struct default_scheduler {
     using result_type = void;
@@ -1487,6 +1487,78 @@ struct main_scheduler {
                 delete f;
             }, new f_t(std::move(f))), 0);
     }
+};
+
+#elif STLAB_TASK_SYSTEM == STLAB_TASK_SYSTEM_WINDOWS
+
+inline FILETIME time_point_to_FILETIME(const std::chrono::system_clock::time_point& tp) {
+    FILETIME ft = { 0, 0 };
+    SYSTEMTIME st = { 0 };
+    time_t t = std::chrono::system_clock::to_time_t(tp);
+    tm utc_tm;
+    if (!gmtime_s(&utc_tm, &t)) {
+        st.wSecond = static_cast<WORD>(utc_tm.tm_sec);
+        st.wMinute = static_cast<WORD>(utc_tm.tm_min);
+        st.wHour = static_cast<WORD>(utc_tm.tm_hour);
+        st.wDay = static_cast<WORD>(utc_tm.tm_mday);
+        st.wMonth = static_cast<WORD>(utc_tm.tm_mon + 1);
+        st.wYear = static_cast<WORD>(utc_tm.tm_year + 1900);
+        st.wMilliseconds = std::chrono::duration_cast<std::chrono::milliseconds>(tp.time_since_epoch()).count() % 1000;
+        SystemTimeToFileTime(&st, &ft);
+    }
+    return ft;
+}
+
+
+class task_system
+{
+public:
+    template <typename F>
+    void async_(F&& f) {
+        auto work = CreateThreadpoolWork(&callback_impl<F>,
+            new F(std::forward<F>(f)),
+            nullptr);
+        if (work == nullptr) {
+            throw std::bad_alloc();
+        }
+        SubmitThreadpoolWork(work);
+    }
+
+    template <typename F>
+    void async_(std::chrono::system_clock::time_point time_point, F&& f) {
+
+        auto timer = CreateThreadpoolTimer(&timer_callback_impl<F>,
+            new F(std::forward<F>(f)),
+            nullptr);
+        if (timer == nullptr) {
+            throw std::bad_alloc();
+        }
+
+        auto file_time = time_point_to_FILETIME(time_point);
+
+        SetThreadpoolTimer(timer,
+            &file_time,
+            0,
+            0);
+    }
+private:
+
+    template <typename F>
+    static void CALLBACK callback_impl(PTP_CALLBACK_INSTANCE /*instance*/,
+        PVOID                 parameter,
+        PTP_WORK              /*Work*/) {
+        std::unique_ptr<F> f(static_cast<F*>(parameter));
+        (*f)();
+    }
+
+    template <typename F>
+    static void CALLBACK timer_callback_impl(PTP_CALLBACK_INSTANCE /*Instance*/,
+        PVOID                 parameter,
+        PTP_TIMER             /*timer*/) {
+        std::unique_ptr<F> f(static_cast<F*>(parameter));
+        (*f)();
+    }
+
 };
 
 #elif STLAB_TASK_SYSTEM == STLAB_TASK_SYSTEM_PORTABLE
